# packages/contracts/src/homebrew_mlflow/contracts/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Homebrew MLflow API",
    "version": "0.1.0"
  },
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/runs/{run_id}/attachments": {
      "post": {
        "operationId": "uploadRunAttachment", "summary": "Upload a policy-permitted small Run attachment", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["path", "file"],
                "properties": {
                  "path": {"type": "string"},
                  "file": {"type": "string", "format": "binary"}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "Attachment stored", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AttachmentFile"}}}}, "409": {"description": "Path already contains different content", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      },
      "get": {
        "operationId": "listRunAttachments", "summary": "List Run attachments", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "path", "in": "query", "required": false, "schema": {"type": "string", "default": ""}}],
        "responses": {"200": {"description": "Attachment entries", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AttachmentListResponse"}}}}}
      }
    },
    "/api/v1/runs/{run_id}/attachments/content": {
      "get": {
        "operationId": "downloadRunAttachment", "summary": "Download one authorized Run attachment", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "path", "in": "query", "required": true, "schema": {"type": "string", "minLength": 1}}],
        "responses": {"200": {"description": "Attachment bytes", "content": {"application/octet-stream": {"schema": {"type": "string", "format": "binary"}}}}, "404": {"description": "Attachment not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
...
      "AttachmentFile": {
        "type": "object", "additionalProperties": false, "required": ["path", "is_dir", "file_size"],
        "properties": {"path": {"type": "string"}, "is_dir": {"type": "boolean"}, "file_size": {"type": ["integer", "null"], "minimum": 0}}
      },
      "AttachmentListResponse": {
        "type": "object", "additionalProperties": false, "required": ["files"],
        "properties": {"files": {"type": "array", "items": {"$ref": "#/components/schemas/AttachmentFile"}}}
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id", "details", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "request_id": {"type": "string"},
          "details": {"type": "object"},
          "retryable": {"type": "boolean"}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/Error"}}
      }
    }
  }
}